requirements.txt
# Core LLM & RAG Frameworks
openai
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core
chromadb>=0.5.0
sentence-transformers>=2.6.0
# llama-cpp-python>=0.2.84
huggingface_hub>=0.23.0
torch
streamlit

# ML / Data Science Utilities
scikit-learn>=1.5.0
pandas>=2.2.0
tqdm>=4.66.0
tabulate
pydantic>=2.0
langdetect
numpy
matplotlib

# Backend / API Layer
fastapi>=0.110.0
uvicorn>=0.30.0
pytest>=7.4.0

# Document Processing & OCR
docling>=1.16.0
easyocr
pillow
pytesseract
pymupdf
pymupdf4llm
pdfplumber
tiktoken
camelot-py
camelot-py[cv]
PyPDF2>=3.0.1

# Data Versioning & Validation
dvc
great_expectations
gcsfs

# Airflow Orchestration
apache-airflow==3.0.0
apache-airflow-providers-sqlite
apache-airflow-providers-http
apache-airflow-providers-postgres

# Experiment Tracking
wandb

# Speech & Audio (STT, TTS, STS)
gTTS
pygame
sounddevice
scipy
faster-whisper
groq
python-dotenv
edge-tts
simpleaudio
pydub

# Additional Utilities
python-multipart
requests
dvc[gs]
pytest 
pytest-asyncio 
httpx

# Agent & LangGraph Dependencies (NEW)
langgraph==0.2.28
python-dateutil==2.9.0